+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Summary of main functional changes.
 
 ## Unreleased
 
+## 26.06.1
+
 - Prevent article aside to be above dropdown menu.
 - Skip invalid articles in feeds.
 - Use article URL as a fallback title if slugify fails.